+++ R/generics.R
@@ -23,6 +23,11 @@ calc_nqx <- function(mod, ...) UseMethod("calc_nqx")
 pop15to49 <- function(mod, ...) UseMethod("pop15to49")
 artpop15to49 <- function(mod, ...) UseMethod("artpop15to49")
 artpop15plus <- function(mod, ...) UseMethod("artpop15plus")
+
+#' @export
 artcov15to49 <- function(mod, ...) UseMethod("artcov15to49")
+
+#' @export
 artcov15plus <- function(mod, ...) UseMethod("artcov15plus")
+
 age15pop <- function(mod, ...) UseMethod("age15pop")